+++ client/apps/messenger/src/features/room/components/Attachments.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Box, IconButton, Typography } from "@mui/material";
 import CloseIcon from "@mui/icons-material/Close";
 import { useParams } from "react-router-dom";
-import AttachmentManager from "../../lib/AttachmentManager";
-import getFileIcon from "../../lib/getFileIcon";
+import AttachmentManager from "../lib/AttachmentManager";
+import getFileIcon from "../lib/getFileIcon";
 
 type AttachmentsProps = {
     files: File[];
