+++ client/apps/messenger/src/features/room/components/emojiPicker/EmojiImg.tsx
