+++ client/apps/messenger/src/features/room/components/emojiPicker/data.ts
