+++ client/apps/messenger/src/features/room/components/emojiPicker/icons/ActivityIcon.tsx
