+++ client/apps/messenger/src/features/room/components/emojiPicker/icons/FlagsIcon.tsx
