+++ client/apps/messenger/src/features/room/components/emojiPicker/icons/FoodIcon.tsx
