+++ client/apps/messenger/src/features/room/components/emojiPicker/icons/NatureIcon.tsx
