+++ client/apps/messenger/src/features/room/components/emojiPicker/index.tsx
