+++ client/apps/messenger/src/features/room/components/emojiPicker/utils/getEmojis.ts
